Rework examples for more narrative results and continue to iterate on copy.
Previously, the "largest" bridge was a big dumb car bridge that nobody
would want to walk along. Instead, lets get a list of all the nice
wissahickon creek bridges within the wissahickon valley park, which
would all be pretty nice to walk along.

Also: include a photo and a bit of narrative around the lovely thomas
mill covered bridge.

Also: added a section about rust "attributes" (#[foo]). Unfortunately
now `attributes` is an overloaded term in this article, which frequently
uses the term "attributes" to mean "the non-geometry data portion of a
feature". Maybe it's OK in context? Or maybe we should double down on
the use of another term like "properties" (which is what geojson calls
them). QGIS calls them "attributes"