unfortunate code fix (tests passing again).
If we want to deserialize to a geojson::Feature, we need to keep all the
type/foreign members etc.

However, if we want to deserialize as a "simple feature" - prinicipally
dropping all fields other than geometry and mapping each member of "properties"
to a top-level struct fields - we need a different visitor. I can't
think of a way to do this generically, so we have a separate flavor of
method.

This is really unfortunate though because it means the user has to think
about "Do I want to deserialize a geojson::Feature or some other kind of
entity" and then pick the right method. The compiler won't help them -
if they choose the wrong thing it'll fail at run time.