add favicon
It seems like the only way to override the favicon is to have a copy in
your src root, which will then be copied into the site directory,
overriding whatever the theme default is.

Notably, there doesn't seem to be a way to specify another favicon (like
a favicon.ico) without forking the theme.